Eliminate all references to stdenv.lib, add tests that build for all backends

+++ nixproc/backends/bsdrc/rcsubr.nix
@@ -1,12 +1,15 @@
 {stdenv, forceDisableUserChange}:
 
-stdenv.mkDerivation {
-  name = "rc.subr";
-  src = /etc/rc.subr;
+if forceDisableUserChange then
   # Disable the limits command when we want to deploy processes as an unprivileged user
-  buildCommand = if forceDisableUserChange then ''
-    sed -e 's|limits -C $_login_class $_limits||' $src > $out
-  '' else ''
-    cp $src $out
-  '';
-}
+  stdenv.mkDerivation {
+    name = "rc.subr";
+    src = /etc/rc.subr;
+
+    buildCommand = ''
+      sed -e 's|limits -C $_login_class $_limits||' $src > $out
+    '';
+  }
+else
+  # Otherwise, simply return the path to the rc subroutines
+  "/etc/rc.subr"
